fix-msvc-type-trait-warnings

+++ include/jwt-cpp/jwt.h
@@ -2507,13 +2507,13 @@ namespace jwt {
 
 		template<typename string_type, typename integer_type>
 		using is_substr_start_end_index_signature =
-			typename std::is_same<decltype(std::declval<string_type>().substr(std::declval<integer_type>(),
-																			  std::declval<integer_type>())),
+			typename std::is_same<decltype(std::declval<string_type>().substr(static_cast<size_t>(std::declval<integer_type>()),
+																			  static_cast<size_t>(std::declval<integer_type>()))),
 								  string_type>;
 
 		template<typename string_type, typename integer_type>
 		using is_substr_start_index_signature =
-			typename std::is_same<decltype(std::declval<string_type>().substr(std::declval<integer_type>())),
+			typename std::is_same<decltype(std::declval<string_type>().substr(static_cast<size_t>(std::declval<integer_type>()))),
 								  string_type>;
 
 		template<typename string_type>
